requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
dlib @ /Scripts/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2023.2
